Update v1_BoostController.ts
the boostpow-js libraries BoostPowJobModel.fromRawTransaction was updated to require the vout of the boost output, defaulting to 0. previously it would check each output to find the boost output. this api needs to allow the vout to be passed in so if the vout is not zero, the job can be submitted.